Removed stdint.h

+++ tbl.c
@@ -19,7 +19,6 @@
  * 3. This notice may not be removed or altered from any source distribution.
 */
 
-#include <stdint.h>
 #include <assert.h>
 #include <string.h>
 #include "tbl.h"
@@ -28,13 +27,13 @@
 #define XXH_NO_STREAM 1
 #include "xxhash.h"
 
-void tbl_init(struct tbl *t, struct tbl_bkt *array, uint16_t n_lg2, uint16_t keylen)
+void tbl_init(struct tbl *t, struct tbl_bkt *array, unsigned short n_lg2, unsigned short keylen)
 {
 	assert(t && array && n_lg2 && keylen);
 	if (t->n)
 		memset(array, 0, sizeof(struct tbl_bkt) << n_lg2);
 	if (!t->seed)
-		t->seed = (uint64_t)t ^ keylen;
+		t->seed = (unsigned long)t ^ keylen;
 	else
 		t->seed++;
 	t->a = array;
@@ -42,16 +41,16 @@ void tbl_init(struct tbl *t, struct tbl_bkt *array, uint16_t n_lg2, uint16_t key
 	t->max = 1 << n_lg2;
 	t->max_lg2 = n_lg2;
 	t->keylen = keylen;
-	t->hashmask = ~(UINT32_MAX << n_lg2);
+	t->hashmask = ~(ULONG_MAX << n_lg2);
 	return;
 }
 
 int tbl_put(struct tbl *t, void *value)
 {
 	assert(t && value);
-	uint32_t hash = (uint32_t)XXH3_64bits_withSeed((char*)value, t->keylen, t->seed);
-	uint32_t pos = hash & t->hashmask;
-	uint32_t off = 0;
+	unsigned int hash = (unsigned int)XXH3_64bits_withSeed((char*)value, t->keylen, t->seed);
+	unsigned int pos = hash & t->hashmask;
+	unsigned int off = 0;
 	if (t->n == t->max)
 		return -1;
 	while (1){
@@ -71,10 +70,10 @@ int tbl_put(struct tbl *t, void *value)
 void *tbl_get(struct tbl *t, const char *key)
 {
 	assert(t && key);
-	uint32_t hash = (uint32_t)XXH3_64bits_withSeed(key, t->keylen, t->seed);
-	uint32_t pos = hash & t->hashmask;
+	unsigned int hash = (unsigned int)XXH3_64bits_withSeed(key, t->keylen, t->seed);
+	unsigned int pos = hash & t->hashmask;
 
-	for (uint32_t off=t->a[pos].maxoff; off >= 0; off--){
+	for (unsigned int off=t->a[pos].maxoff; off >= 0; off--){
 		if (t->a[pos].value && hash == t->a[pos].hash){
 			if (!(memcmp(key, t->a[pos].value, t->keylen)))
 				return t->a[pos].value;
@@ -87,11 +86,11 @@ void *tbl_get(struct tbl *t, const char *key)
 void *tbl_remove(struct tbl *t, const char *key)
 {
 	assert(t && key);
-	uint32_t hash = (uint32_t)XXH3_64bits_withSeed(key, t->keylen, t->seed);
-	uint32_t pos = hash & t->hashmask;
+	unsigned int hash = (unsigned int)XXH3_64bits_withSeed(key, t->keylen, t->seed);
+	unsigned int pos = hash & t->hashmask;
 	void *found;
 
-	for (uint32_t off=t->a[pos].maxoff; off >= 0; off--){
+	for (unsigned int off=t->a[pos].maxoff; off >= 0; off--){
 		if (t->a[pos].value && hash == t->a[pos].hash){
 			if (!(memcmp(key, t->a[pos].value, t->keylen))){
 				found = t->a[pos].value;
@@ -109,7 +108,7 @@ void tbl_copy(struct tbl *dest, struct tbl *src)
 {
 	assert(src && dest);
 	assert(dest->max >= src->max);
-	for (uint32_t i=0; i != src->max; i++){
+	for (unsigned int i=0; i != src->max; i++){
 		if (src->a[i].value)
 			assert(tbl_put(dest, src->a[i].value) == 0);
 	}
